refactor(DatasetPreview): split components into different files
